Move helper method to synchronization utils
